Convert project to Typescript
