+++ .circleci/config.yml
@@ -1,0 +1,117 @@
+version: 2.1
+
+variables:
+  - &working_directory '~/tax-ca'
+
+executors:
+  node:
+    docker:
+      - image: circleci/node:10
+
+commands:
+  attach_project_directory:
+    steps:
+      - attach_workspace:
+          at: *working_directory
+
+  # NodeJS / Yarn
+  restore_node_cache:
+    steps:
+      - restore_cache:
+          keys:
+            - &node_cache 'v1-node-{{ arch }}-{{ checksum "yarn.lock" }}'
+            - 'v1-node-{{ arch }}-'
+  save_node_cache:
+    steps:
+      - save_cache:
+          key: *node_cache
+          paths: 'node_modules'
+
+  restore_yarn_cache:
+    steps:
+      - restore_cache:
+          keys:
+            - &yarn_cache 'v1-yarn-{{ arch }}'
+  save_yarn_cache:
+    steps:
+      - save_cache:
+          key: *yarn_cache
+          paths:
+            - '~/.cache/yarn'
+
+  yarn_cached_install:
+    parameters:
+      yarn_arguments:
+        type: string
+        default: ''
+    steps:
+      - restore_node_cache
+      - restore_yarn_cache
+      - run:
+          name: 'Install Node Dependencies'
+          command: 'yarn --frozen-lockfile << parameters.yarn_arguments >>'
+      - save_node_cache
+      - save_yarn_cache
+
+  yarn_install:
+    steps:
+      - attach_project_directory
+      - yarn_cached_install
+
+  yarn_build:
+    steps:
+      - run: yarn build
+
+jobs:
+  checkout:
+    executor: node
+    working_directory: *working_directory
+    steps:
+      - checkout
+      - run:
+          name: 'Save initial build number'
+          command: 'echo ${CIRCLE_BUILD_NUM} > .build-number'
+      - persist_to_workspace:
+          root: '.'
+          paths: '.'
+
+  yarn_install:
+    executor: node
+    working_directory: *working_directory
+    steps:
+      - yarn_install
+
+  lint:
+    executor: node
+    working_directory: *working_directory
+    steps:
+      - attach_project_directory
+      - restore_node_cache
+      - run: 'mkdir -p build/tests/lint/'
+      - run: 'yarn lint:ci'
+      - store_test_results:
+          path: 'build/tests'
+
+  build:
+    executor: node
+    working_directory: *working_directory
+    steps:
+      - attach_project_directory
+      - restore_node_cache
+      - yarn_build
+
+workflows:
+  version: 2
+
+  build:
+    jobs:
+      - checkout
+      - yarn_install:
+          requires:
+            - checkout
+      - lint:
+          requires:
+            - yarn_install
+      - build:
+          requires:
+            - yarn_install
