+++ .editorconfig
@@ -1,0 +1,29 @@
+root = true
+
+[*]
+charset = utf-8
+indent_style = space
+indent_size = 4
+max_line_length = 120
+continuation_indent_size = 4
+insert_final_newline = true
+trim_trailing_whitespace = true
+
+[*.sh]
+end_of_line = lf
+
+[*.{json,js,ts}]
+indent_brace_style = K&R
+curly_bracket_next_line = false
+
+[*.{js,ts}]
+quote_style = single
+spaces_around_operators = true
+spaces_around_brackets = both
+
+[*.{json,yml}]
+indent_size = 2
+
+[*.md]
+max_line_length = off
+trim_trailing_whitespace = false
