+++ src/investments/rrsp.ts
@@ -1,0 +1,11 @@
+// tslint:disable:max-line-length
+/*
+Sources:
+	https://www.canada.ca/en/revenue-agency/services/tax/registered-plans-administrators/pspa/mp-rrsp-dpsp-tfsa-limits-ympe.html
+	http://www.rrq.gouv.qc.ca/SiteCollectionDocuments/www.rrq.gouv.qc/Francais/pdf/rcr/annexe06.pdf
+
+Revised 2018-12-21
+*/
+// tslint:enable:max-line-length
+
+export = { MAX_CONTRIBUTION: 26500 };
