--- src/investments/tfsa.js
@@ -1,11 +1,0 @@
-/*
-Sources:
-	https://www.canada.ca/en/revenue-agency/services/tax/registered-plans-administrators/pspa/mp-rrsp-dpsp-tfsa-limits-ympe.html
-	
-Revised 2018-12-21
-*/
-
-export default {
-	MAX_CONTRIBUTION: 6000,
-	ROUNDING_FACTOR: 500
-};
