+++ src/investments/tfsa.ts
@@ -1,9 +1,13 @@
+// tslint:disable:max-line-length
 /*
 Sources:
 	https://www.canada.ca/en/revenue-agency/services/tax/registered-plans-administrators/pspa/mp-rrsp-dpsp-tfsa-limits-ympe.html
-	http://www.rrq.gouv.qc.ca/SiteCollectionDocuments/www.rrq.gouv.qc/Francais/pdf/rcr/annexe06.pdf
 
 Revised 2018-12-21
 */
+// tslint:enable:max-line-length
 
-export default { MAX_CONTRIBUTION: 26500 };
+export = {
+    MAX_CONTRIBUTION: 6000,
+    ROUNDING_FACTOR: 500,
+};
