+++ src/misc/cpi.ts
@@ -1,0 +1,149 @@
+/*
+Sources:
+	http://www.banqueducanada.ca/taux/indices-des-prix/ipc/ (IPC Global)
+
+Revised 2018-12-21
+*/
+
+export = {
+    2009: {
+        JAN: 113.0,
+        FEB: 113.8,
+        MAR: 114.0,
+        APR: 113.9,
+        MAY: 114.7,
+        JUN: 115.1,
+        JUL: 114.7,
+        AUG: 114.7,
+        SEP: 114.7,
+        OCT: 114.6,
+        NOV: 115.2,
+        DEC: 114.8,
+    },
+    2010: {
+        JAN: 115.1,
+        FEB: 115.6,
+        MAR: 115.6,
+        APR: 116.0,
+        MAY: 116.3,
+        JUN: 116.2,
+        JUL: 116.8,
+        AUG: 116.7,
+        SEP: 116.9,
+        OCT: 117.4,
+        NOV: 117.5,
+        DEC: 117.5,
+    },
+    2011: {
+        JAN: 117.8,
+        FEB: 118.1,
+        MAR: 119.4,
+        APR: 119.8,
+        MAY: 120.6,
+        JUN: 119.8,
+        JUL: 120.0,
+        AUG: 120.3,
+        SEP: 120.6,
+        OCT: 120.8,
+        NOV: 120.9,
+        DEC: 120.2,
+    },
+    2012: {
+        JAN: 120.7,
+        FEB: 121.2,
+        MAR: 121.7,
+        APR: 122.2,
+        MAY: 122.1,
+        JUN: 121.6,
+        JUL: 121.5,
+        AUG: 121.8,
+        SEP: 122.0,
+        OCT: 122.2,
+        NOV: 121.9,
+        DEC: 121.2,
+    },
+    2013: {
+        JAN: 121.3,
+        FEB: 122.7,
+        MAR: 122.9,
+        APR: 122.7,
+        MAY: 123.0,
+        JUN: 123.0,
+        JUL: 123.1,
+        AUG: 123.1,
+        SEP: 123.3,
+        OCT: 123.0,
+        NOV: 123.0,
+        DEC: 122.7,
+    },
+    2014: {
+        JAN: 123.1,
+        FEB: 124.1,
+        MAR: 124.8,
+        APR: 125.2,
+        MAY: 125.8,
+        JUN: 125.9,
+        JUL: 125.7,
+        AUG: 125.7,
+        SEP: 125.8,
+        OCT: 125.9,
+        NOV: 125.4,
+        DEC: 124.5,
+    },
+    2015: {
+        JAN: 124.3,
+        FEB: 125.4,
+        MAR: 126.3,
+        APR: 126.2,
+        MAY: 126.9,
+        JUN: 127.2,
+        JUL: 127.3,
+        AUG: 127.3,
+        SEP: 127.1,
+        OCT: 127.2,
+        NOV: 127.1,
+        DEC: 126.5,
+    },
+    2016: {
+        JAN: 126.8,
+        FEB: 127.1,
+        MAR: 127.9,
+        APR: 128.3,
+        MAY: 128.8,
+        JUN: 129.1,
+        JUL: 128.9,
+        AUG: 128.7,
+        SEP: 128.8,
+        OCT: 129.1,
+        NOV: 128.6,
+        DEC: 128.4,
+    },
+    2017: {
+        JAN: 129.5,
+        FEB: 129.7,
+        MAR: 129.9,
+        APR: 130.4,
+        MAY: 130.5,
+        JUN: 130.4,
+        JUL: 130.4,
+        AUG: 130.5,
+        SEP: 130.8,
+        OCT: 130.9,
+        NOV: 131.3,
+        DEC: 130.8,
+    },
+    2018: {
+        JAN: 131.7,
+        FEB: 132.5,
+        MAR: 132.9,
+        APR: 133.3,
+        MAY: 133.4,
+        JUN: 133.6,
+        JUL: 134.3,
+        AUG: 134.2,
+        SEP: 134.3,
+        OCT: 134.1,
+        NOV: 133.5,
+        DEC: 133.5,
+    },
+};
