--- src/misc/iqpf.js
@@ -1,20 +1,0 @@
-/*
-Sources:
-	https://www.iqpf.org/docs/default-source/outils/iqpf-normes-projection2019.pdf
-	http://app.iqpf.org/guidelines
-
-Revised 2019-05-01
-*/
-
-export default {
-	INFLATION: 0.021,
-	RETURN_RATES: {
-		SHORT_TERM: 0.03,
-		FIXED_INCOME: 0.039,
-		CANADIAN_EQUITIES: 0.061,
-		CONSERVATIVE_PORTFOLIO: 0.032,
-		BALANCED_PORTFOLIO: 0.037,
-		DYNAMIC_PORTFOLIO: 0.045,
-	},
-	BORROWING_RATE: 0.05,
-};
