+++ src/pension/cpp.ts
@@ -13,7 +13,66 @@ Sources:
 Revised 2019-01-17
 */
 
-export default {
+const MAX_INCOME: { [K: number]: number } = {
+    1966: 5000,
+    1967: 5000,
+    1968: 5100,
+    1969: 5200,
+    1970: 5300,
+    1971: 5400,
+    1972: 5500,
+    1973: 5900,
+    1974: 6600,
+    1975: 7400,
+    1976: 8300,
+    1977: 9300,
+    1978: 10400,
+    1979: 11700,
+    1980: 13100,
+    1981: 14700,
+    1982: 16500,
+    1983: 18500,
+    1984: 20800,
+    1985: 23400,
+    1986: 25800,
+    1987: 25900,
+    1988: 26500,
+    1989: 27700,
+    1990: 28900,
+    1991: 30500,
+    1992: 32200,
+    1993: 33400,
+    1994: 34400,
+    1995: 34900,
+    1996: 35400,
+    1997: 35800,
+    1998: 36900,
+    1999: 37400,
+    2000: 37600,
+    2001: 38300,
+    2002: 39100,
+    2003: 39900,
+    2004: 40500,
+    2005: 41100,
+    2006: 42100,
+    2007: 43700,
+    2008: 44900,
+    2009: 46300,
+    2010: 47200,
+    2011: 48300,
+    2012: 50100,
+    2013: 51100,
+    2014: 52500,
+    2015: 53600,
+    2016: 54900,
+    2017: 55300,
+    2018: 55900,
+    2019: 57400,
+};
+
+type MaxIncomeByYear = keyof typeof MAX_INCOME;
+
+export = {
     CONTRIBUTIONS: {
         MAX_PENSIONABLE_EARNINGS: 57400,
         MIN_PENSIONABLE_EARNINGS: 3500,
@@ -35,9 +94,9 @@ export default {
         FROM_45_TO_64: 7519.56,
         OVER_64_WITHOUT_PENSION: 8313,
     },
-    getAAF(age) {
-        const lower = 0.0060; const
-            higher = 0.0070;
+    getAAF(age: number): number {
+        const lower = 0.0060;
+        const higher = 0.0070;
 
         if (age < 60) {
             return 0;
@@ -50,20 +109,23 @@ export default {
         }
         return (1 + ((Math.min(70, age) - 65) * 12 * higher));
     },
-    getAverageIndexationRate() {
-        let sum = 0;
+    getAverageIndexationRate(): string {
+        const sum = this.INDEXATION_RATES_REFERENCES.reduce((previous, current) => previous + current[1], 0);
+        const averageAsString: string = (sum / this.INDEXATION_RATES_REFERENCES.length).toString();
 
-        for (let i = 0; i < this.INDEXATION_RATES_REFERENCES.length; i++) {
-            sum += this.INDEXATION_RATES_REFERENCES[i][1];
-        }
-
-        return Number.parseFloat(sum / this.INDEXATION_RATES_REFERENCES.length).toPrecision(3);
+        return Number.parseFloat(averageAsString).toPrecision(3);
     },
-    getMPEA(year) {
-        return (this.MAX_INCOME[year - 4] + this.MAX_INCOME[year - 3] + this.MAX_INCOME[year - 2] + this.MAX_INCOME[year - 1] + this.MAX_INCOME[year]) / 5;
+    getMPEA(year: MaxIncomeByYear): number {
+        return (this.MAX_INCOME[year - 4] + this.MAX_INCOME[year - 3] + this.MAX_INCOME[year - 2]
+            + this.MAX_INCOME[year - 1] + this.MAX_INCOME[year]) / 5;
     },
-    getPostRetirementBenefit(pensionnableEarning, ympe, mpea, aaf) {
-        return (pensionnableEarning / ympe) * 0.00625 * mpea * aaf / 12;
+    getPostRetirementBenefit(
+        pensionableEarning: number,
+        yearMaximumPe: number,
+        maximumPeAverage: number,
+        aaf: number,
+    ): number {
+        return (pensionableEarning / yearMaximumPe) * 0.00625 * maximumPeAverage * aaf / 12;
     },
     INDEXATION_RATES_REFERENCES: [ // Previous year inflation used as indexation
         [2007, 0.020],
@@ -86,62 +148,7 @@ export default {
         SURVIVOR_UNDER_65: 7519.56,
         DEATH_BENEFIT: 2500,
     },
-    MAX_INCOME: {
-        1966: 5000,
-        1967: 5000,
-        1968: 5100,
-        1969: 5200,
-        1970: 5300,
-        1971: 5400,
-        1972: 5500,
-        1973: 5900,
-        1974: 6600,
-        1975: 7400,
-        1976: 8300,
-        1977: 9300,
-        1978: 10400,
-        1979: 11700,
-        1980: 13100,
-        1981: 14700,
-        1982: 16500,
-        1983: 18500,
-        1984: 20800,
-        1985: 23400,
-        1986: 25800,
-        1987: 25900,
-        1988: 26500,
-        1989: 27700,
-        1990: 28900,
-        1991: 30500,
-        1992: 32200,
-        1993: 33400,
-        1994: 34400,
-        1995: 34900,
-        1996: 35400,
-        1997: 35800,
-        1998: 36900,
-        1999: 37400,
-        2000: 37600,
-        2001: 38300,
-        2002: 39100,
-        2003: 39900,
-        2004: 40500,
-        2005: 41100,
-        2006: 42100,
-        2007: 43700,
-        2008: 44900,
-        2009: 46300,
-        2010: 47200,
-        2011: 48300,
-        2012: 50100,
-        2013: 51100,
-        2014: 52500,
-        2015: 53600,
-        2016: 54900,
-        2017: 55300,
-        2018: 55900,
-        2019: 57400,
-    },
+    MAX_INCOME,
     MAX_REQUEST_AGE: 70,
     MIN_REQUEST_AGE: 60,
     REPLACEMENT_FACTOR: 0.25,
