--- src/pension/spp.js
@@ -1,8 +1,0 @@
-/*
-Revised 2018-12-21
-*/
-
-export default {
-	MAX_BRIDGE_BENEFIT_AGE: 65,
-	MIN_AGE: 55
-};
