+++ src/pension/spp.ts
@@ -1,0 +1,8 @@
+/*
+Revised 2018-12-21
+*/
+
+export = {
+    MAX_BRIDGE_BENEFIT_AGE: 65,
+    MIN_AGE: 55,
+};
