+++ src/taxes/ei.ts
@@ -1,3 +1,4 @@
+// tslint:disable:max-line-length
 /*
 Sources:
 	https://www.canada.ca/en/employment-social-development/programs/ei/ei-list/ei-employers/2019-maximum-insurable-earnings.html
@@ -5,12 +6,12 @@ Sources:
 
 Revised 2018-12-21
 */
+// tslint:enable:max-line-length
 
-export default {
-	MAX_INSURABLE_EARNINGS: 53100,
-	PREMIUM_RATES: {
-		CA: 0.0162,
-		QC: 0.0125,
-	},
+export = {
+    MAX_INSURABLE_EARNINGS: 53100,
+    PREMIUM_RATES: {
+        CA: 0.0162,
+        QC: 0.0125,
+    },
 };
-
