+++ src/taxes/qpip.ts
@@ -2,14 +2,14 @@
 Sources:
 	https://www.revenuquebec.ca/en/press-room/tax-news/details/161040/2018-12-13/
 	http://www.rqap.gouv.qc.ca/a_propos_regime/information_generale/cotisations_en.asp
-	
+
 Revised 2018-12-21
 */
 
-export default {
-	MAX_INSURABLE_EARNINGS: 76500,
-	PREMIUM_RATES: {
-		SELF_EMPLOYED: 0.00934,
-		SALARIED: 0.00526,
-	},
+export = {
+    MAX_INSURABLE_EARNINGS: 76500,
+    PREMIUM_RATES: {
+        SELF_EMPLOYED: 0.00934,
+        SALARIED: 0.00526,
+    },
 };
