+++ src/utils/collections.ts
@@ -1,0 +1,8 @@
+export function maxBy<T>(
+    values: T[],
+    mapFn: ((item: T) => number),
+): T | undefined {
+    const max = Math.max(...values.map(mapFn));
+
+    return values.find(item => mapFn(item) === max);
+}
