+++ tslint.json
@@ -1,0 +1,11 @@
+{
+  "extends": [
+    "@equisoft/tslint-config"
+  ],
+  "rules": {
+    "linebreak-style": [false],
+    "quotemark": {
+      "options": ["single"]
+    }
+  }
+}
